post to SCM updates
