+++ ICL.DWH.Backend.Core/Services/IPurchaseOrderService.cs
@@ -13,7 +13,7 @@ namespace ICL.DWH.Backend.Core.Services
         IEnumerable<PurchaseOrder> GetPurchaseOrders();
         void UpdatePurchaseOrderByScmId(string bookingId, Guid scmId);
         void UpdatePurchaseOrderAsFailed(string bookingId, string errorMessage);
-        string ValidatePurchaseOrder(string bookingId);
-        Task<string> PostToSCMAsync(string mySbMsg, string bookingNo);
+        string ValidatePurchaseOrder(PurchaseOrder purchaseOrder);
+        PurchaseOrder UpdatePurchaseOrder(PurchaseOrder purchaseOrder);
     }
 }
